Move `write_graphviz_results` from `results.rs` to `graphviz.rs`.
It's more graphviz-y than it is results-y. This lets us reduce
visibility of several types in `graphviz.rs`.
